requirements.txt
python-dotenv
psycopg2-binary
requests
numpy
pandas
tqdm
openai
gradio
streamlit
langchain
tiktoken
faiss-cpu
langchain-openai
langchain_experimental
langchain_chroma
langchain[docarray]
datasets
sentencepiece
matplotlib
google.generativeai
anthropic
unstructured
chromadb
plotly
jupyter-dash
beautifulsoup4
pydub
modal
ollama
accelerate
sentencepiece
bitsandbytes